% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyDiagnostics.R
\docType{package}
\name{StudyDiagnostics-package}
\alias{StudyDiagnostics}
\alias{StudyDiagnostics-package}
\title{StudyDiagnostics: Generating Cohort Diagnostics for Darwin EU studies}
\description{
Generates cohort diagnostics for the cohort definitions of DARWIN EU studies.
}
\keyword{internal}
